+++ Sources/CvTeam.h
@@ -89,7 +89,7 @@ public:
 	int getHasCorporationCount(CorporationTypes eCorporation) const; // Exposed to Python
 
 	int processedTeamCulture() const;
-	unsigned long long countTotalCulture() const; // Exposed to Python
+	uint64_t countTotalCulture() const; // Exposed to Python
 
 	int countNumUnitsByArea(const CvArea* pArea) const; // Exposed to Python
 	int countNumCitiesByArea(const CvArea* pArea) const; // Exposed to Python
